+++ task_set_arginfo.h
@@ -9,10 +9,10 @@
 #define arginfo_class_Async_TaskSet_joinAny         arginfo_class_Async_TaskGroup_any
 #define arginfo_class_Async_TaskSet_joinAll         arginfo_class_Async_TaskGroup_all
 #define arginfo_class_Async_TaskSet_cancel          arginfo_class_Async_TaskGroup_cancel
-#define arginfo_class_Async_TaskSet_seal            arginfo_class_Async_TaskGroup_seal
+#define arginfo_class_Async_TaskSet_close           arginfo_class_Async_TaskGroup_close
 #define arginfo_class_Async_TaskSet_dispose         arginfo_class_Async_TaskGroup_dispose
 #define arginfo_class_Async_TaskSet_isFinished      arginfo_class_Async_TaskGroup_isFinished
-#define arginfo_class_Async_TaskSet_isSealed        arginfo_class_Async_TaskGroup_isSealed
+#define arginfo_class_Async_TaskSet_isClosed        arginfo_class_Async_TaskGroup_isClosed
 #define arginfo_class_Async_TaskSet_count           arginfo_class_Async_TaskGroup_count
 #define arginfo_class_Async_TaskSet_awaitCompletion arginfo_class_Async_TaskGroup_awaitCompletion
 #define arginfo_class_Async_TaskSet_finally         arginfo_class_Async_TaskGroup_finally
@@ -27,10 +27,10 @@ static const zend_function_entry class_Async_TaskSet_methods[] = {
 	ZEND_MALIAS(Async_TaskGroup, joinAny, any, arginfo_class_Async_TaskSet_joinAny, ZEND_ACC_PUBLIC)
 	ZEND_MALIAS(Async_TaskGroup, joinAll, all, arginfo_class_Async_TaskSet_joinAll, ZEND_ACC_PUBLIC)
 	ZEND_MALIAS(Async_TaskGroup, cancel, cancel, arginfo_class_Async_TaskSet_cancel, ZEND_ACC_PUBLIC)
-	ZEND_MALIAS(Async_TaskGroup, seal, seal, arginfo_class_Async_TaskSet_seal, ZEND_ACC_PUBLIC)
+	ZEND_MALIAS(Async_TaskGroup, close, close, arginfo_class_Async_TaskSet_close, ZEND_ACC_PUBLIC)
 	ZEND_MALIAS(Async_TaskGroup, dispose, dispose, arginfo_class_Async_TaskSet_dispose, ZEND_ACC_PUBLIC)
 	ZEND_MALIAS(Async_TaskGroup, isFinished, isFinished, arginfo_class_Async_TaskSet_isFinished, ZEND_ACC_PUBLIC)
-	ZEND_MALIAS(Async_TaskGroup, isSealed, isSealed, arginfo_class_Async_TaskSet_isSealed, ZEND_ACC_PUBLIC)
+	ZEND_MALIAS(Async_TaskGroup, isClosed, isClosed, arginfo_class_Async_TaskSet_isClosed, ZEND_ACC_PUBLIC)
 	ZEND_MALIAS(Async_TaskGroup, count, count, arginfo_class_Async_TaskSet_count, ZEND_ACC_PUBLIC)
 	ZEND_MALIAS(Async_TaskGroup, awaitCompletion, awaitCompletion, arginfo_class_Async_TaskSet_awaitCompletion, ZEND_ACC_PUBLIC)
 	ZEND_MALIAS(Async_TaskGroup, finally, finally, arginfo_class_Async_TaskSet_finally, ZEND_ACC_PUBLIC)